Phase 10a: Fix return-type (45) and narrowing (30) warnings
Eliminate 75 compiler warnings:
- 45 -Wreturn-type: Add missing return statements to fix undefined behavior
- 30 -Wnarrowing: Add explicit static_casts for narrowing conversions

Files: 28 changed across AssemblyUtils, Blurmaps, ColorTable2, DataManager,
DynamicAdaptiveGrid, F2DockClient, fast-LJ, HLevelSet, ImposterRenderer,
InterfaceStats, LBIE_lib, MolSurfAPI, OpenGL_Viewer, PDBParser,
SignDistanceFunction_v2, VolMagick, VolumeLibrary

+++ src/ColorTable2/Table.cpp
@@ -725,9 +725,9 @@ namespace CVCColorTable
   {
      if( !_dirtyContourTree ) return;
   
-     int dim[3] = { _contourVolume.XDim(),
-                    _contourVolume.YDim(),
-                    _contourVolume.ZDim() };
+     int dim[3] = { static_cast<int>(_contourVolume.XDim()),
+                    static_cast<int>(_contourVolume.YDim()),
+                    static_cast<int>(_contourVolume.ZDim()) };
 
      _contourTreeVertices.clear();
      _contourTreeEdges.clear();
@@ -794,9 +794,9 @@ namespace CVCColorTable
      Signature	*sig;
      int dim[3] = 
      {
-        _contourVolume.XDim(),
-        _contourVolume.YDim(),
-        _contourVolume.ZDim()
+        static_cast<int>(_contourVolume.XDim()),
+        static_cast<int>(_contourVolume.YDim()),
+        static_cast<int>(_contourVolume.ZDim())
      };
 
      orig[0] = _contourVolume.XMin();
